frontend/requirements.txt
# المكتبات الأساسية للواجهة
PyQt5>=5.15.4
PyQt5-sip>=12.9.0
PyQt5-stubs>=5.15.2.0
PyQt5-Qt5>=5.15.2

# مكتبات الشبكة والاتصال
requests>=2.26.0
python-socketio>=5.4.0
urllib3>=1.26.7

# مكتبات الصوت للدردشة الصوتية
pyaudio>=0.2.11
numpy>=1.21.0
wave>=0.0.2

# معالجة البيانات والتصميم
Pillow>=8.3.2
python-dotenv>=0.19.0

# أدوات التطوير
flake8>=4.0.1
black>=21.9b0
isort>=5.9.3 